requirements.txt
baidu-aip>=4.16.10
requests>=2.31.0
chardet>=5.0.0
jieba>=0.42.1
python-dotenv>=1.0.1

